Combine the "with" interfaces in v1/partial.
Remove the imageCore dependency from all of them and bias towards narrower interfaces.
