+++ v1/partial/BUILD.bazel
@@ -7,10 +7,7 @@ go_library(
         "doc.go",
         "image.go",
         "uncompressed.go",
-        "with_blob.go",
-        "with_configfile.go",
-        "with_manifest.go",
-        "with_uncompressedlayer.go",
+        "with.go",
     ],
     importpath = "github.com/google/go-containerregistry/v1/partial",
     visibility = ["//visibility:public"],
